Move dlcTest to using testkitcore
