+++ testkit-core/src/main/scala/org/bitcoins/testkitcore/util/BytesUtil.scala
@@ -1,4 +1,4 @@
-package org.bitcoins.testkit.util
+package org.bitcoins.testkitcore.util
 
 import org.bitcoins.core.protocol.dlc.{CETSignatures, FundingSignatures}
 import org.bitcoins.core.protocol.script.{
